My submitted solution for the Hypersonic competition.
This was the last version before I started over from scratch, ie this is the code that I started in the Wood leagues and iterated upon through the Bronze league. This solution is Roomba-esque: it mainly operates turn-by-turn and doesn't do much prediction of future game states. I don't think there's much more value to be squeezed out of this one.